# lab2/openapi.yaml
openapi: 3.0.0
info:
  title: Social Network API
  version: 1.0.0
  description: API for social network services including user management, wall posts, and chat
servers:
  - url: http://localhost:8000
    description: API Service
  - url: http://localhost:8001
    description: Chat Service
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Message:
      type: object
      properties:
        sender:
          type: string
        receiver:
          type: string
        content:
          type: string
        timestamp:
          type: string
          format: date-time
paths:
...
  /api/v1/message/send:
    post:
      summary: Send message
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Message'
      responses:
        '200':
          description: Message sent successfully
  /api/v1/message:
    get:
      summary: Get messages
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Message' 